Upgrade V8 to 7.6.53

+++ v8/src/flags.h
@@ -40,10 +40,10 @@ class V8_EXPORT_PRIVATE FlagList {
   // The following syntax for flags is accepted (both '-' and '--' are ok):
   //
   //   --flag        (bool flags only)
-  //   --noflag      (bool flags only)
+  //   --no-flag     (bool flags only)
   //   --flag=value  (non-bool flags only, no spaces around '=')
   //   --flag value  (non-bool flags only)
-  //   --            (equivalent to --js_arguments, captures all remaining args)
+  //   --            (capture all remaining args in JavaScript)
   static int SetFlagsFromCommandLine(int* argc,
                                      char** argv,
                                      bool remove_flags);
@@ -51,7 +51,7 @@ class V8_EXPORT_PRIVATE FlagList {
   // Set the flag values by parsing the string str. Splits string into argc
   // substrings argv[], each of which consisting of non-white-space chars,
   // and then calls SetFlagsFromCommandLine() and returns its result.
-  static int SetFlagsFromString(const char* str, int len);
+  static int SetFlagsFromString(const char* str, size_t len);
 
   // Reset all flags to their default value.
   static void ResetAllFlags();
